requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
openai>=1.30.0
sarvamai>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.1